pyproject.toml
[project]
name = "rag-project-2026"
version = "0.1.0"
description ="Création d'une solution RAG dockeriser permetttant la lecture de fichier volumineux (Doc Civil, Code général des impôts)"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-community>=0.4.0",
    "langchain-text-splitters>=1.0.0",
    "chromadb>=1.0.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=5.0.0",
    "streamlit>=1.40.0",
    "langchain-mistralai>=1.1.2",
    "python-dotenv>=1.2.2",
]

[tool.uv]
dev-dependencies = [
    "jupyter",
    "notebook"
]
